MVVM for AppStartup
