--- SigfolioWallet.UWP/AppStart.xaml
@@ -1,45 +1,0 @@
-﻿<views:MvxWindowsPage
-    xmlns:views="using:MvvmCross.Platforms.Uap.Views"
-    x:Class="SigfolioWallet.AppStart"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:SigfolioWallet"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:controls="using:Microsoft.Toolkit.Uwp.UI.Controls"
-
-    mc:Ignorable="d" Background="{ThemeResource SystemControlAcrylicWindowBrush}">
-
-    <StackPanel VerticalAlignment="Center">
-        <controls:DropShadowPanel Height="670" Width="570" BlurRadius="40" ShadowOpacity=".2" OffsetX="0" OffsetY="-8"  >
-            <StackPanel Orientation="Vertical">
-                <!--<Button HorizontalAlignment="Left" Name="btnBack" Margin="0 0 0 -42"  Background="{ThemeResource SystemControlBackgroundAccentBrush}" FocusVisualSecondaryBrush="{ThemeResource SystemControlBackgroundAccentBrush}" Foreground="#fff" VerticalAlignment="Top" RenderTransformOrigin="0.547,1.467" Click="Button_Click" Width="48" Height="42" >
-                    <Button.Resources>
-                        <ResourceDictionary>
-                            <ResourceDictionary.ThemeDictionaries>
-                                <ResourceDictionary x:Key="Dark">
-                                    <SolidColorBrush x:Key="ButtonBackgroundPointerOver" Color="{ThemeResource SystemAccentColor}"/>
-                                </ResourceDictionary>
-                                <ResourceDictionary x:Key="Light">
-                                    <SolidColorBrush x:Key="ButtonBackgroundPointerOver" Color="{ThemeResource SystemAccentColor}"/>
-                                </ResourceDictionary>
-                            </ResourceDictionary.ThemeDictionaries>
-                        </ResourceDictionary>
-                    </Button.Resources>
-                    <StackPanel Orientation="Horizontal">
-                        <FontIcon Glyph="&#xE72B;" FontSize="14"></FontIcon>
-
-                    </StackPanel>
-                </Button>-->
-                <Frame x:Name="RootFrame">
-                </Frame>
-            </StackPanel>
-        </controls:DropShadowPanel>
-        <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Padding="10" Width="380">
-            <FontIcon FontSize="22" Glyph="&#xE72E;"></FontIcon>
-            <TextBlock Padding="12" TextWrapping="WrapWholeWords" FontSize="12" Width="330">To keep your account safe, your private key is encrypted locally on your device.</TextBlock>
-
-        </StackPanel>
-    </StackPanel>
-
-</views:MvxWindowsPage>
