+++ SigfolioWallet.UWP/Views/AppStartup/GettingStarted.xaml
@@ -1,10 +1,11 @@
-﻿<Page
-    x:Class="SigfolioWallet.Views.GettingStarted"
+﻿<views:MvxWindowsPage
+    x:Class="SigfolioWallet.Views.AppStartup.GettingStarted"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:SigfolioWallet.Views"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:views="using:MvvmCross.Platforms.Uap.Views"
     mc:Ignorable="d"
     Width="570" Height="650"
     Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
@@ -35,7 +36,7 @@
 
         </StackPanel>
         <ProgressBar VerticalAlignment="Bottom" Margin="0 0 0 90" Width="550" Value="10" ></ProgressBar>
-        <Button HorizontalAlignment="Left" Background="{ThemeResource SystemControlBackgroundAccentBrush}" Foreground="#fff" Margin="445,0,0,20" VerticalAlignment="Bottom" RenderTransformOrigin="0.547,1.467" Width="115" Content="Next"/>
+        <Button HorizontalAlignment="Left" Background="{ThemeResource SystemControlBackgroundAccentBrush}" Command="{Binding Next}" Foreground="#fff" Margin="445,0,0,20" VerticalAlignment="Bottom" RenderTransformOrigin="0.547,1.467" Width="115" Content="Next"/>
         <Button HorizontalAlignment="Right" Background="{ThemeResource SystemControlBackgroundAccentBrush}" Foreground="#fff"  VerticalAlignment="Bottom" RenderTransformOrigin="0.547,1.467" Width="120" Content="Import Existing" Margin="10,0,440,20"/>
     </Grid>
-</Page>
+</views:MvxWindowsPage>
