+++ SigfolioWallet.UWP/Views/AppStartup/GettingStarted.xaml.cs
@@ -1,0 +1,27 @@
+﻿using Windows.UI.Xaml.Controls;
+using MvvmCross.Platforms.Uap.Presenters.Attributes;
+using MvvmCross.Platforms.Uap.Views;
+using MvvmCross.ViewModels;
+using SigfolioWallet.Core.ViewModels;
+using SigfolioWallet.Core.ViewModels.AppStartup;
+
+// The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
+
+namespace SigfolioWallet.Views.AppStartup
+{
+    /// <summary>
+    /// An empty page that can be used on its own or navigated to within a Frame.
+    /// </summary>
+    [MvxRegionPresentation("AppStartupFrame")]
+    [MvxViewFor(typeof(GettingStartedViewModel))]
+    public sealed partial class GettingStarted : MvxWindowsPage
+    {
+        public GettingStarted()
+        {
+            this.InitializeComponent();
+        }
+
+        public new GettingStartedViewModel ViewModel => (GettingStartedViewModel)base.ViewModel;
+
+    }
+}
