--- SigfolioWallet.UWP/Views/GettingStarted.xaml.cs
@@ -1,37 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Navigation;
-
-// The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
-
-namespace SigfolioWallet.Views
-{
-    /// <summary>
-    /// An empty page that can be used on its own or navigated to within a Frame.
-    /// </summary>
-    public sealed partial class GettingStarted : Page
-    {
-        public GettingStarted()
-        {
-            this.InitializeComponent();
-
-            //cbAccounts.DataContext = ViewModel;
-
-            //if (ViewModel.Wallet.CurrentAccountId == 0)
-            //{
-            //    cbAccounts.SelectedValue = ViewModel.Wallet.CurrentAccountId;
-            //}
-        }
-    }
-}
