+++ SigfolioWallet.UWP/Views/HistoryView.xaml
